1. Broke our huge MBStore file into 3 files, MBStore, MBArticlesStore, and MBDevotionsStore 2. Removed the year date component from devotion.dateComponentsForNotification so they will work indefinitely
