+++ iOS Client/AppDelegate/AppDelegate.swift
@@ -13,6 +13,7 @@ import CoreData
 class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
     var appCoordinator: AppCoordinator!
+    let articlesStore = MBArticlesStore()
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         self.window = UIWindow(frame: UIScreen.main.bounds)
@@ -57,7 +58,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             return
         }
         
-        MBStore().syncAllData(persistentContainer: persistentContainer) { (isNewData: Bool?, syncErr: Error?) in
+        articlesStore.syncAllData(persistentContainer: persistentContainer) { (isNewData: Bool?, syncErr: Error?) in
             if syncErr != nil {
                 completionHandler(.failed)
             } else {
